requirements.txt
anthropic
pandas
openpyxl
streamlit
python-dotenv
pyinstaller
